+++ src/libjasper/jp2/jp2_cod.c
@@ -271,7 +271,7 @@ jp2_box_t *jp2_box_get(jas_stream_t *in)
 		if (jp2_getuint64(in, &extlen)) {
 			goto error;
 		}
-		box->len = extlen;
+		box->len = (uint_fast32_t)extlen; // JMW?
 		box->data_len = box->len - 8;
 	}
 	if (box->len != 0 && box->len < JP2_BOX_HDRLEN) {
@@ -715,8 +715,8 @@ static int jp2_putuint32(jas_stream_t *out, uint_fast32_t val)
 
 static int jp2_putuint64(jas_stream_t *out, uint_fast64_t val)
 {
-	if (jp2_putuint32(out, (val >> 32) & 0xffffffffUL) ||
-	  jp2_putuint32(out, val & 0xffffffffUL)) {
+	if (jp2_putuint32(out, (uint_fast32_t)((val >> 32) & 0xffffffffUL)) ||
+	  jp2_putuint32(out, (uint_fast32_t)(val & 0xffffffffUL))) {
 		return -1;
 	}
 	return 0;
